extract_model {sleuth}	R Documentation

Extract a model from a sleuth object

Description

This function extracts the parameter estimates from a sleuth model after it has been fit with sleuth_fit.

Usage

extract_model(obj, which_model)

Arguments

obj
a sleuth object.

which_model
a model fitted with sleuth_fit.

Value

a data frame including a column for the target_id, the term (which coefficient), and the corresponding standard error.

[Package sleuth version 0.30.0 Index]