gene_from_gene {sleuth}	R Documentation

Get the gene ID using other gene identifiers

Description

Get the target_id of a gene using other gene identifiers. The identifiers found under the obj$gene_column are often difficult to remember (e.g. ensembl gene ID, ENSG00000111640). This function allows a user to find that difficult-to-remember identifier using more-easily-remembered identifiers, such as gene symbol (e.g. "GAPDH").

Usage

gene_from_gene(obj, gene_colname, gene_name)

Arguments

obj
a sleuth object

gene_colname
the name of the column containing 'gene_name'. This parameter refers to the name of the column that the gene you are searching for appears in. Check the column names using colnames(obj$target_mapping).

gene_name
a string containing the name of the gene you are interested in.

Value

a character vector containing the target_id of the gene, found under obj$gene_column within obj$target_mapping. If the column name provided is the same as obj$gene_column, and the gene_name used is found, that gene_name will be returned.

Examples

  ## Not run: gene_from_gene(obj, "gene_symbol", "GAPDH")

[Package sleuth version 0.30.0 Index]