sleuth_live_settings {sleuth}	R Documentation

settings for sleuth_live

Description

This is a helper function for setting preferences in sleuth live. Currently, this is somewhat limited, but it will be expanded in the future.

Usage

sleuth_live_settings(test_type = "wt")

Arguments

test_type
either 'wt' for the wald test or 'lrt' for the likelihood ratio test

Value

a named to list with the options and settings

[Package sleuth version 0.30.0 Index]