sleuth_lrt {sleuth}	R Documentation

sleuth likelihood ratio test

Description

compute the likelihood ratio test for 2 models. this requires that the null model be nested in the alternate model

Usage

sleuth_lrt(obj, null_model, alt_model)

Arguments

obj
a sleuth object

null_model
the null (or "reduced") model

alt_model
the alternate (or "full") model

Value

an updated sleuth object with a likelihood ratio test computation

See Also

models to view which models have been fit and which coefficients can be tested, sleuth_results to get back a data.frame of the results

[Package sleuth version 0.30.0 Index]